Added C++ implementation and R wrapper for new covariance priors
New priors are block diagonal inverse-wishart, where the blocks are determined by whether the nodes are high or low in the tree.  The code seems to be working, though the prior values need a bit of experimentation before it's ready.  Currently, we seem to be experiencing an issue more than 99% of the sequencing reads accumulate in the same subcommunity.